refactor and extract layout component
