+++ src/app/index.tsx
@@ -1,44 +1,17 @@
-import NDK from '@nostr-dev-kit/ndk';
-import NDKCacheAdapterDexie from '@nostr-dev-kit/ndk-cache-dexie';
-import { useAutoLogin, useNostrHooks } from 'nostr-hooks';
-import { useEffect, useMemo } from 'react';
-import { Outlet, useParams } from 'react-router-dom';
+import { RouterProvider } from 'react-router-dom';
 
-import { useGlobalNdk } from '@/shared/hooks';
+import { router } from '@/app/router';
+
+import { ThemeProvider } from '@/shared/components/theme-provider';
 
 import './index.css';
 
 export const App = () => {
-  const { relay } = useParams();
-
-  const ndk = useMemo(
-    () =>
-      new NDK({
-        explicitRelayUrls: relay ? [relay] : undefined,
-        autoConnectUserRelays: false,
-        autoFetchUserMutelist: false,
-        cacheAdapter: relay
-          ? new NDKCacheAdapterDexie({
-              dbName: `db-${relay}`,
-            })
-          : undefined,
-      }),
-    [relay],
+  return (
+    <>
+      <ThemeProvider defaultTheme="dark" storageKey="vite-ui-theme">
+        <RouterProvider router={router} />
+      </ThemeProvider>
+    </>
   );
-
-  useNostrHooks(ndk);
-
-  const { globalNdk, setGlobalSigner } = useGlobalNdk();
-
-  useEffect(() => {
-    globalNdk.connect();
-  }, [globalNdk]);
-
-  useAutoLogin();
-
-  useEffect(() => {
-    setGlobalSigner(ndk.signer);
-  }, [ndk.signer, setGlobalSigner]);
-
-  return <Outlet />;
 };
